Refactor MakeKOSKSginature move it to secrets
Use the stored secret everywhere
